allow clicking anywhere on a card in the creation view
