+++ src/simplified/baseCard.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {Card, Row} from "react-bootstrap";
+import {Card, Container, Row} from "react-bootstrap";
 import {useState, useEffect} from "react";
 import { CSSTransition } from 'react-transition-group';
 import State from "../state";
@@ -18,7 +18,8 @@ interface BaseCardProps {
     profile?: string,
     wait?: number,
     content?: JSX.Element,
-    clickable?: boolean
+    onClick?: () => void,
+    infoLink?: string,
     link?: string
     bsBg?: string,
     bsText?: Color | string,
@@ -34,8 +35,6 @@ interface BaseCardProps {
 export const BaseCard = (props: BaseCardProps) => {
     const [show, setShow] = useState(false);
     
-    const navigate = useNavigate();
-    
     useEffect(() => {
         const timeoutId = setTimeout(() => {
             setShow(true);
@@ -47,7 +46,6 @@ export const BaseCard = (props: BaseCardProps) => {
     const content = props.content;
     let headerPadding = {};
     if (props.title == "") headerPadding = {padding:"7px"};
-    const resourceType = (props.header == "Questionnaire")? "Questionnaire":friendlyToFhir(props.header);
 
     return (
         <CSSTransition
@@ -60,80 +58,33 @@ export const BaseCard = (props: BaseCardProps) => {
             bg={props.bsBg}
             text={props.bsText as Color}
             border={props.bsBorder}
-            onClick={(e: React.MouseEvent<HTMLDivElement, MouseEvent>) => {
-                if (props.disabled) {
-                    return;
-                }
-                if (e.target instanceof Element && e.target.tagName !== "svg" && e.target.tagName !== "path" && props.clickable && resourceType!='') {
-                    setShow(false);
-                    if (State.get().bundle?.resources.length) {
-                        State.get().bundle.set("pos", State.get().bundle.resources.length-1);
-                        State.get().ui.set("openMode", "insert");
-                    }
-                    const nextId = incrementNextId(); // Saving some trouble by using this -- we should decide on a standard way to generate unique URLs
-                    const { referencedResourceName, referencedResourceUrl } = generateResourceReference(resourceType, nextId);
-                    const json = {
-                        resourceType: "Bundle",
-                        entry: [
-                            {
-                                resource: {
-                                    resourceType: resourceType,
-                                    name: referencedResourceName,
-                                    url: referencedResourceUrl,
-                                    meta: {profile: [props.profile]}
-                                }
-                            },
-                            {
-                                resource: {
-                                    resourceType: PLAN_DEFINITION,
-                                    library: "", // r4 expects library as an array, cpg expects a single value (we are always using the cpg spec in basic view)
-                                    action: [
-                                        {
-                                            title: "",
-                                            description: "",
-                                            condition: [],
-                                            definitionCanonical: referencedResourceUrl
-                                        }
-                                    ],
-                                    relatedArtifact: []
-                                }
-                            }
-                        ]
-                    };
-                    State.emit("load_json_resource", json);
-                    // Set current editor position to the last resource (should be the PlanDefinition in `json` after the "load_json_resource" call)
-                    State.emit("set_bundle_pos", State.get().bundle.resources.length-1);
-                    if (props.title === "An Example Questionaire") { // Fixes https://github.com/PuraJuniper/SAGE/pull/84#issuecomment-1104155086
-                                                                     // We should be able to remove this If with SAGE-332
-                        navigate(`/author?next=edit/${State.get().bundle.resources.length-1}`);
-                    }
-                    else {
-                        navigate(`/edit/${State.get().bundle.resources.length-1}`);
-                    }
-                }
-                if(props.title == 'Create Cards'){
-                    CreateCardWorkflow(navigate)
-                }
-                if(props.title == 'View Cards'){
-                    navigate('/view-cards')
-                }
-            }}
+            onClick={props.onClick}
         >
             <Card.Header style={headerPadding} hidden = {props.hideHeader}>
                 {props.header}
             </Card.Header>
-            <Card.Body >
-                <div>
-                    <span style={{ fontSize: "20px", textAlign: "right" }}>
-                        <a href={props.profile} target="_blank" rel="noreferrer" className="c-tooltip">
-                            <FontAwesomeIcon icon={faInfoCircle} style={{'color':props.IconColor}} />
-                            <span className="c-tooltiptext">FHIR Docs</span>
-                        </a>
-                    </span>
-                </div>
-                <Card.Title style={{ fontSize: props.titleSize, textAlign: "center" }}>{props.title}</Card.Title>
-                <Card.Text>{content}</Card.Text>
-                <FontAwesomeIcon icon={props.cardImage} style={{'color':props.IconColor, 'height':props.IconSize}} />
+            <Card.Body>
+                <Container>
+                    <Row className="align-content-end">
+                        {props.infoLink !== undefined ?
+                            <span style={{ fontSize: "20px", textAlign: "right" }}>
+                                <a href={props.infoLink} target="_blank" rel="noreferrer" className="c-tooltip">
+                                    <FontAwesomeIcon icon={faInfoCircle} style={{'color':props.IconColor}} />
+                                    <span className="c-tooltiptext">FHIR Docs</span>
+                                </a>
+                            </span> :
+                            <div className="m-3" />}
+                    </Row>
+                    <Row className="align-content-center">
+                        <Card.Title className="col" style={{ fontSize: props.titleSize, textAlign: "center" }}>{props.title}</Card.Title>
+                    </Row>
+                    <Row className="align-content-center">
+                        <Card.Text className="col">{content}</Card.Text>
+                    </Row>
+                    <Row className="align-content-center">
+                        <FontAwesomeIcon className="col pe-none" icon={props.cardImage} style={{'color':props.IconColor, 'height':props.IconSize}} />
+                    </Row>
+                </Container>
             </Card.Body>
         </Card>
         </CSSTransition>
