+++ src/simplified/selectView.tsx
@@ -2,10 +2,11 @@ import { faUserDoctor } from '@fortawesome/pro-solid-svg-icons';
 import React from 'react';
 import { Col, Container, Row } from "react-bootstrap";
 import { NavigateFunction, useNavigate } from "react-router-dom";
+import { generateResourceReference, incrementNextId } from '../helpers/schema-utils';
 import State from "../state";
 import { BaseCard } from "./baseCard";
 import { AUTHOR_THEN_CARD_ROUTE, AUTHOR_THEN_EXIT_ROUTE } from './basicView';
-import { friendlyResourceRoot } from "./nameHelpers";
+import { ACTIVITY_DEFINITION, friendlyResourceRoot, PLAN_DEFINITION } from "./nameHelpers";
 import { Progress } from './topProgressBar';
 
 
@@ -52,7 +53,47 @@ const SelectView = () => {
                                                     title={resource.FRIENDLY}
                                                     hideHeader = {true}
                                                     wait={i * 25}
-                                                    clickable={true}
+                                                    infoLink={resource.DEFAULT_PROFILE_URI}
+                                                    onClick={() => {
+                                                        if (State.get().bundle?.resources.length) {
+                                                            State.get().bundle.set("pos", State.get().bundle.resources.length-1);
+                                                            State.get().ui.set("openMode", "insert");
+                                                        }
+                                                        const nextId = incrementNextId(); // Saving some trouble by using this -- we should decide on a standard way to generate unique URLs
+                                                        const { referencedResourceName, referencedResourceUrl } = generateResourceReference(ACTIVITY_DEFINITION, nextId);
+                                                        const json = {
+                                                            resourceType: "Bundle",
+                                                            entry: [
+                                                                {
+                                                                    resource: {
+                                                                        resourceType: ACTIVITY_DEFINITION,
+                                                                        name: referencedResourceName,
+                                                                        url: referencedResourceUrl,
+                                                                        meta: {profile: [resource.DEFAULT_PROFILE_URI]}
+                                                                    }
+                                                                },
+                                                                {
+                                                                    resource: {
+                                                                        resourceType: PLAN_DEFINITION,
+                                                                        library: "", // r4 expects library as an array, cpg expects a single value (we are always using the cpg spec in basic view)
+                                                                        action: [
+                                                                            {
+                                                                                title: "",
+                                                                                description: "",
+                                                                                condition: [],
+                                                                                definitionCanonical: referencedResourceUrl
+                                                                            }
+                                                                        ],
+                                                                        relatedArtifact: []
+                                                                    }
+                                                                }
+                                                            ]
+                                                        };
+                                                        State.emit("load_json_resource", json);
+                                                        // Set current editor position to the last resource (should be the PlanDefinition in `json` after the "load_json_resource" call)
+                                                        State.emit("set_bundle_pos", State.get().bundle.resources.length-1);
+                                                        navigate(`/edit/${State.get().bundle.resources.length-1}`);
+                                                    }}
                                                     profile={resource.DEFAULT_PROFILE_URI}
                                                     titleSize='15px'
                                                     IconSize= '50px'
